+++ qiskit_aqua/svm_classical/svm_classical_binary.py
@@ -35,8 +35,8 @@ class SVM_Classical_Binary(SVM_Classical_ABC):
         """
         train the svm
         Args:
-            training_input: dictionary which maps each class to the points in the class
-            class_labels: array of classes. For example: ['A', 'B']
+            test_input (dict): dictionary which maps each class to the points in the class
+            class_labels (list): list of classes. For example: ['A', 'B']
         """
         training_points, training_points_labels, label_to_class = get_points_and_labels(training_input, class_labels)
 
@@ -64,8 +64,8 @@ class SVM_Classical_Binary(SVM_Classical_ABC):
         """
         test the svm
         Args:
-            test_input: dictionary which maps each class to the points in the class
-            class_labels: array of classes. For example: ['A', 'B']
+            test_input (dict): dictionary which maps each class to the points in the class
+            class_labels (list): list of classes. For example: ['A', 'B']
         """
         test_points, test_points_labels, label_to_labelclass = get_points_and_labels(test_input, class_labels)
 
@@ -108,9 +108,8 @@ class SVM_Classical_Binary(SVM_Classical_ABC):
         """
         predict using the svm
         Args:
-            test_points: the points (array)
+            test_points (numpy.ndarray): the points
         """
-
         alphas = self._ret['svm']['alphas']
         bias = self._ret['svm']['bias']
         SVMs = self._ret['svm']['support_vectors']
