+++ qiskit_aqua/svm_qkernel/cost_helpers.py
@@ -15,8 +15,9 @@
 # limitations under the License.
 # =============================================================================
 
-from collections import Counter
 import numpy as np
+
+from collections import Counter
 
 
 def assign_label(key, class_labels):
