+++ qiskit_aqua/svm_qkernel/qpsolver.py
@@ -16,7 +16,9 @@
 # =============================================================================
 
 import logging
+
 import numpy as np
+
 from cvxopt import matrix, solvers
 
 logger = logging.getLogger(__name__)
