+++ test/test_svm_classical.py
@@ -15,13 +15,12 @@
 # limitations under the License.
 # =============================================================================
 
-from qiskit_aqua.utils.multiclass.data_preprocess import *
-from qiskit_aqua.svm_qkernel.data_preprocess import *
-
-
 import numpy as np
 
 from test.common import QiskitAquaTestCase
+
+from qiskit_aqua.utils.multiclass.data_preprocess import *
+from qiskit_aqua.svm_qkernel.data_preprocess import *
 from qiskit_aqua import run_algorithm, get_algorithm_instance
 from qiskit_aqua.input import get_input_instance
