+++ test/test_svm_qkernel.py
@@ -15,13 +15,12 @@
 # limitations under the License.
 # =============================================================================
 
-from qiskit_aqua.utils.multiclass.data_preprocess import *
-from qiskit_aqua.svm_qkernel.data_preprocess import *
-
-
 import numpy as np
 
 from test.common import QiskitAquaTestCase
+
+from qiskit_aqua.utils.multiclass.data_preprocess import *
+from qiskit_aqua.svm_qkernel.data_preprocess import *
 from qiskit_aqua import run_algorithm, get_algorithm_instance
 from qiskit_aqua.input import get_input_instance
 
@@ -30,6 +29,33 @@ class TestSVMQKernel(QiskitAquaTestCase):
 
     def setUp(self):
         self.random_seed = 10598
+        self.training_data = {'A': np.asarray([[2.95309709, 2.51327412], [3.14159265, 4.08407045]]),
+                              'B': np.asarray([[4.08407045, 2.26194671], [4.46106157, 2.38761042]])}
+        self.testing_data = {'A': np.asarray([[3.83274304, 2.45044227]]),
+                             'B': np.asarray([[3.89557489, 0.31415927]])}
+
+        self.ref_kernel_matrix_training = np.asarray([[1., 0.84277344, 0.12109375, 0.37011719],
+                                                      [0.84277344, 1., 0.10742188, 0.44042969],
+                                                      [0.12109375, 0.10742188, 1., 0.6484375],
+                                                      [0.37011719, 0.44042969, 0.6484375, 1.]])
+
+        self.ref_kernel_matrix_testing = np.asarray([[0.12988281, 0.15820312, 0.45996094, 0.14648438],
+                                                     [0.33300781, 0.38085938, 0.01660156, 0.1484375]])
+
+        self.ref_support_vectors = np.asarray([[2.95309709, 2.51327412],
+                                               [3.14159265, 4.08407045],
+                                               [4.08407045, 2.26194671],
+                                               [4.46106157, 2.38761042]])
+        self.ref_alpha = np.asarray([[0.53596828],
+                                     [1.29460379],
+                                     [0.11233349],
+                                     [1.71823858]])
+
+        self.ref_bias = np.asarray([0.02252818])
+
+        self.svm_input = get_input_instance('SVMInput')
+        self.svm_input.training_dataset = self.training_data
+        self.svm_input.test_dataset = self.testing_data
 
 
     def test_classical_binary(self):
@@ -198,3 +224,58 @@ class TestSVMQKernel(QiskitAquaTestCase):
         print(result)
         self.assertEqual(result['test_success_ratio'], 0.5555555555555556)
         self.assertEqual(result['predicted_labels'], ['A', 'A', 'C', 'A', 'A', 'A', 'C', 'C', 'C'])
+
+    def test_svm_qkernel_via_run_algorithm(self):
+
+        params = {
+            'problem': {'name': 'svm_classification', 'random_seed': self.random_seed},
+            'algorithm': {'name': 'SVM_QKernel'},
+            'backend': {'name': 'local_qasm_simulator_py', 'shots': 1024}
+        }
+        result = run_algorithm(params, self.svm_input)
+
+        np.testing.assert_array_almost_equal(
+            result['kernel_matrix_training'], self.ref_kernel_matrix_training, decimal=4)
+        np.testing.assert_array_almost_equal(
+            result['kernel_matrix_testing'], self.ref_kernel_matrix_testing, decimal=4)
+
+        self.assertEqual(len(result['svm']['support_vectors']), 4)
+        np.testing.assert_array_almost_equal(
+            result['svm']['support_vectors'], self.ref_support_vectors, decimal=4)
+
+        np.testing.assert_array_almost_equal(result['svm']['alphas'], self.ref_alpha, decimal=4)
+        np.testing.assert_array_almost_equal(result['svm']['bias'], self.ref_bias, decimal=4)
+
+        self.assertEqual(result['test_success_ratio'], 0.0)
+
+    def test_svm_qkernel_directly(self):
+        svm = get_algorithm_instance("SVM_QKernel")
+        svm.setup_quantum_backend(backend='local_qasm_simulator_py', shots=1024)
+        svm.random_seed = self.random_seed
+
+        params = {
+            'problem': {'name': 'svm_classification', 'random_seed': self.random_seed},
+            'algorithm': {'name': 'SVM_QKernel'},
+            'backend': {'name': 'local_qasm_simulator_py', 'shots': 1024}
+        }
+
+        svm.init_params(params, self.svm_input)
+        result = svm.run()
+
+        np.testing.assert_array_almost_equal(
+            result['kernel_matrix_training'], self.ref_kernel_matrix_training, decimal=4)
+        np.testing.assert_array_almost_equal(
+            result['kernel_matrix_testing'], self.ref_kernel_matrix_testing, decimal=4)
+
+        self.assertEqual(len(result['svm']['support_vectors']), 4)
+        np.testing.assert_array_almost_equal(
+            result['svm']['support_vectors'], self.ref_support_vectors, decimal=4)
+
+        np.testing.assert_array_almost_equal(result['svm']['alphas'], self.ref_alpha, decimal=4)
+        np.testing.assert_array_almost_equal(result['svm']['bias'], self.ref_bias, decimal=4)
+
+
+        self.assertEqual(result['test_success_ratio'], 0.0)
+
+
+
